fix add filter by title in the UI
