+++ frontend/src/constants/actionTypes.js
@@ -38,4 +38,3 @@ export const PROFILE_FAVORITES_PAGE_UNLOADED =
 export const PROFILE_FAVORITES_PAGE_LOADED = "PROFILE_FAVORITES_PAGE_LOADED";
 export const UPDATE_SEARCH_FIELD = "UPDATE_FIELD_EDITOR";
 export const QUERY_SUBMITTED = "QUERY_SUBMITTED";
-
